refactor + support custom scale models
